requirements.txt
asgiref==3.2.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
python-decouple==3.4
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
django-jsonfield==1.4.0
gunicorn==20.0.4
idna==2.9
jwcrypto==0.7
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
authorizenet==1.1.3
Pillow==7.1.2